requirements.txt
# Core ML & Data Manipulation
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.1
lightgbm>=4.1.0
joblib>=1.3.2

# Request limits & Web scraping (Feature extraction)
requests>=2.31.0
beautifulsoup4>=4.12.2

# Web Framework (API server)
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
jinja2>=3.1.2

# Environment variables
python-dotenv>=1.0.0

# Testing & ML Utils
pytest>=7.4.2
httpx>=0.25.0
shap>=0.45.0
